style(msvc): fix C4244/C4267 in OptionRow/JsonUtil/GraphDisplay/AdjustSync/ActorScroller/DynamicActorScroller/Actor (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- OptionRow.cpp (4): a size_t course-entry count into int; a
  std::min<unsigned int> call's second arg; a lua_pushnumber(size_t)
  accessor.
- JsonUtil.h (4): four near-identical root.resize(v.size()) sites
  across SerializeVectorObjects/SerializeVectorPointers (x2)/
  SerializeArrayValues template helpers, each needing
  Json::Value::ArrayIndex instead of size_t.
- GraphDisplay.cpp (4): a size_t vertex count into DrawQuads's int; a
  size_t/int division into a fan count; two theme-metric ints into
  float size members.
- AdjustSync.cpp (4): a size_t size snapshotted before/after a filter
  call into an int counter (both sides of the diff); a size_t into
  unsigned int; a FormatNumberAndSuffix(int) call fed a size_t loop
  index.
- ActorScroller.cpp (4) + DynamicActorScroller.cpp (2, bonus find
  while in the file): a size_t sub-actor count into int; two
  std::ceil(float) results into int; three wrap(int&, size_t) sites
  across both scroller files.
- Actor.cpp (4): two RString::Left calls fed size_t-minus arithmetic
  parsing command/message names; a uint64_t timer value into a
  generic_global_timer_update(float, ...) call; a
  lua_pushnumber(size_t) accessor.

+++ src/AdjustSync.cpp
@@ -250,9 +250,9 @@ void AdjustSync::AutosyncTempo()
 		 * However, throwing away the ones with high error should be enough
 		 * in most cases. */
 		float fFilteredError = 0;
-		s_iStepsFiltered = s_vAutosyncTempoData.size();
+		s_iStepsFiltered = static_cast<int>(s_vAutosyncTempoData.size());
 		FilterHighErrorPoints( s_vAutosyncTempoData, fSlope, fIntercept, ERROR_TOO_HIGH );
-		s_iStepsFiltered -= s_vAutosyncTempoData.size();
+		s_iStepsFiltered -= static_cast<int>(s_vAutosyncTempoData.size());
 
 		if( !CalcLeastSquares( s_vAutosyncTempoData, fSlope, fIntercept, fFilteredError ) )
 			return;
@@ -339,7 +339,7 @@ void AdjustSync::GetSyncChangeTextSong( std::vector<RString> &vsAddTo )
 		LuaHelpers::ReportScriptError("The sync overlay's " #segments_name " segment list is a different size from the song's.  Please report this bug with steps to reproduce it."); \
 	}
 
-		unsigned int iOriginalSize = vsAddTo.size();
+		unsigned int iOriginalSize = static_cast<unsigned int>(vsAddTo.size());
 		TimingData &original = s_vpTimingDataOriginal[0];
 		TimingData &testing = GAMESTATE->m_pCurSong->m_SongTiming;
 
@@ -376,7 +376,7 @@ void AdjustSync::GetSyncChangeTextSong( std::vector<RString> &vsAddTo )
 			}
 
 			RString s = ssprintf( TEMPO_SEGMENT_FROM.GetValue(),
-					FormatNumberAndSuffix(i+1).c_str(), fOld, fNew );
+					FormatNumberAndSuffix(static_cast<int>(i+1)).c_str(), fOld, fNew );
 
 			vsAddTo.push_back( s );
 		}
